requirements.txt
discord.py>=2.3.2
asyncpg>=0.29.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
